+++ cloudbot/clients/irc.py
@@ -11,20 +11,22 @@ from typing import Mapping, Optional
 
 from irclib.parser import Message
 
-from cloudbot.client import Client, client, ClientConnectError
+from cloudbot.client import Client, ClientConnectError, client
 from cloudbot.event import Event, EventType, IrcOutEvent
 from cloudbot.util import async_util
 
 logger = logging.getLogger("cloudbot")
 
-irc_nick_re = re.compile(r'[A-Za-z0-9^{\}\[\]\-`_|\\]+')
+irc_nick_re = re.compile(r"[A-Za-z0-9^{\}\[\]\-`_|\\]+")
 
-irc_bad_chars = ''.join([chr(x) for x in list(range(0, 1)) + list(range(4, 32)) + list(range(127, 160))])
-irc_clean_re = re.compile('[{}]'.format(re.escape(irc_bad_chars)))
+irc_bad_chars = "".join(
+    [chr(x) for x in list(range(0, 1)) + list(range(4, 32)) + list(range(127, 160))]
+)
+irc_clean_re = re.compile("[{}]".format(re.escape(irc_bad_chars)))
 
 
 def irc_clean(dirty):
-    return irc_clean_re.sub('', dirty)
+    return irc_clean_re.sub("", dirty)
 
 
 irc_command_to_event_type = {
@@ -32,7 +34,7 @@ irc_command_to_event_type = {
     "JOIN": EventType.join,
     "PART": EventType.part,
     "KICK": EventType.kick,
-    "NOTICE": EventType.notice
+    "NOTICE": EventType.notice,
 }
 
 content_params = {
@@ -57,6 +59,14 @@ chan_params = {
     "353": 2,
     "366": 1,
     "324": 1,
+    "329": 1,
+    "332": 1,
+    "333": 1,
+    # WHOIS
+    "310": 1,
+    "311": 1,
+    "312": 1,
+    "318": 1,
 }
 
 target_params = {
@@ -69,13 +79,17 @@ target_params = {
 def decode(bytestring):
     """
     Tries to decode a bytestring using multiple encoding formats
+
+    >>> decode(bytes([0x80, 0xbf, 0x81]) + '\u200b'.encode())
+    '\u200b'
     """
-    for codec in ('utf-8', 'iso-8859-1', 'shift_jis', 'cp1252'):
+    for codec in ("utf-8", "shift_jis", "cp1252"):
         try:
             return bytestring.decode(codec)
         except UnicodeDecodeError:
             continue
-    return bytestring.decode('utf-8', errors='ignore')
+
+    return bytestring.decode("utf-8", errors="ignore")
 
 
 def _get_param(msg: Message, index_map: Mapping[str, int]) -> Optional[str]:
@@ -108,16 +122,16 @@ class IrcClient(Client):
         super().__init__(bot, _type, name, nick, channels=channels, config=config)
 
         self.target_nick = nick
-        conn_config = config['connection']
-        self.use_ssl = conn_config.get('ssl', False)
-        self._ignore_cert_errors = conn_config.get('ignore_cert', False)
-        self._timeout = conn_config.get('timeout', 300)
-        self.server = conn_config['server']
-        self.port = conn_config.get('port', 6667)
+        conn_config = config["connection"]
+        self.use_ssl = conn_config.get("ssl", False)
+        self._ignore_cert_errors = conn_config.get("ignore_cert", False)
+        self._timeout = conn_config.get("timeout", 300)
+        self.server = conn_config["server"]
+        self.port = conn_config.get("port", 6667)
 
         local_bind = (
-            conn_config.get('bind_addr', False),
-            conn_config.get('bind_port', 0),
+            conn_config.get("bind_addr", False),
+            conn_config.get("bind_port", 0),
         )
         if local_bind[0] is False:
             local_bind = False
@@ -135,7 +149,7 @@ class IrcClient(Client):
     def make_ssl_context(self, conn_config):
         if self.use_ssl:
             ssl_context = ssl.create_default_context()
-            client_cert = conn_config.get('client_cert')
+            client_cert = conn_config.get("client_cert")
             if client_cert:
                 path = Path(client_cert)
                 if path.exists():
@@ -177,12 +191,17 @@ class IrcClient(Client):
             try:
                 await self.connect(self._timeout)
             except (TimeoutError, asyncio.TimeoutError):
-                logger.error("[%s] Timeout occurred while connecting to %s", self.name, self.describe_server())
+                logger.error(
+                    "[%s] Timeout occurred while connecting to %s",
+                    self.name,
+                    self.describe_server(),
+                )
             except (socket.error, socket.gaierror, OSError, ssl.SSLError):
                 logger.error(
                     "[%s] Error occurred while connecting to %s (%s)",
-                    self.name, self.describe_server(),
-                    traceback.format_exc().splitlines()[-1]
+                    self.name,
+                    self.describe_server(),
+                    traceback.format_exc().splitlines()[-1],
                 )
             except Exception as e:
                 raise ClientConnectError(self.name, self.describe_server()) from e
@@ -192,10 +211,8 @@ class IrcClient(Client):
             sleep_time = random.randrange(self._timeout)
             canceller = asyncio.shield(self.cancelled_future)
             try:
-                await asyncio.wait_for(
-                    canceller, timeout=sleep_time
-                )
-            except asyncio.CancelledError:
+                await asyncio.wait_for(canceller, timeout=sleep_time)
+            except (asyncio.CancelledError, asyncio.TimeoutError):
                 pass
 
     async def connect(self, timeout=None):
@@ -203,7 +220,9 @@ class IrcClient(Client):
         Connects to the IRC server, or reconnects if already connected.
         """
         if self._connecting:
-            raise ValueError("Attempted to connect while another connect attempt is happening")
+            raise ValueError(
+                "Attempted to connect while another connect attempt is happening"
+            )
 
         self._connecting = True
         try:
@@ -226,7 +245,11 @@ class IrcClient(Client):
             optional_params["local_addr"] = self.local_bind
 
         coro = self.loop.create_connection(
-            partial(_IrcProtocol, self), host=self.server, port=self.port, ssl=self.ssl_context, **optional_params
+            partial(_IrcProtocol, self),
+            host=self.server,
+            port=self.port,
+            ssl=self.ssl_context,
+            **optional_params,
         )
 
         if timeout is not None:
@@ -235,7 +258,9 @@ class IrcClient(Client):
         self._transport, self._protocol = await coro
 
         tasks = [
-            self.bot.plugin_manager.launch(hook, Event(bot=self.bot, conn=self, hook=hook))
+            self.bot.plugin_manager.launch(
+                hook, Event(bot=self.bot, conn=self, hook=hook)
+            )
             for hook in self.bot.plugin_manager.connect_hooks
             if not hook.clients or self.type in hook.clients
         ]
@@ -421,9 +446,13 @@ class _IrcProtocol(asyncio.Protocol):
                 bot=self.bot, hook=out_sieve, conn=self.conn, irc_raw=line
             )
 
-            ok, new_line = await self.bot.plugin_manager.internal_launch(out_sieve, event)
+            ok, new_line = await self.bot.plugin_manager.internal_launch(
+                out_sieve, event
+            )
             if not ok:
-                logger.warning("Error occurred in outgoing sieve, falling back to old behavior")
+                logger.warning(
+                    "Error occurred in outgoing sieve, falling back to old behavior"
+                )
                 logger.debug("Line was: %s", line)
                 filtered = False
                 break
@@ -461,7 +490,9 @@ class _IrcProtocol(asyncio.Protocol):
             except Exception:
                 logger.exception(
                     "[%s] Error occurred while parsing IRC line '%s' from %s",
-                    self.conn.name, line, self.conn.describe_server()
+                    self.conn.name,
+                    line,
+                    self.conn.describe_server(),
                 )
             else:
                 # handle the message, async
@@ -485,22 +516,20 @@ class _IrcProtocol(asyncio.Protocol):
             content = None
 
         # Event type
-        event_type = irc_command_to_event_type.get(
-            command, EventType.other
-        )
+        event_type = irc_command_to_event_type.get(command, EventType.other)
         target = _get_param(message, target_params)
 
         # Parse for CTCP
         if event_type is EventType.message and content_raw.startswith("\x01"):
             possible_ctcp = content_raw[1:]
-            if content_raw.endswith('\x01'):
+            if content_raw.endswith("\x01"):
                 possible_ctcp = possible_ctcp[:-1]
 
-            if '\x01' in possible_ctcp:
+            if "\x01" in possible_ctcp:
                 logger.debug(
                     "[%s] Invalid CTCP message received, "
                     "treating it as a mornal message",
-                    self.conn.name
+                    self.conn.name,
                 )
                 ctcp_text = None
             else:
@@ -518,6 +547,7 @@ class _IrcProtocol(asyncio.Protocol):
 
         # Channel
         channel = _get_param(message, chan_params)
+
         prefix = message.prefix
         if prefix is None:
             nick = None
@@ -539,13 +569,29 @@ class _IrcProtocol(asyncio.Protocol):
             # Channel for a PM is the sending user
             if channel == self.conn.nick.lower():
                 channel = nick.lower()
+        else:
+            # If the channel isn't set, it's the sending user/server
+            channel = nick.lower() if nick else nick
 
         # Set up parsed message
         # TODO: Do we really want to send the raw `prefix` and `command_params` here?
         event = Event(
-            bot=self.bot, conn=self.conn, event_type=event_type, content_raw=content_raw, content=content,
-            target=target, channel=channel, nick=nick, user=user, host=host, mask=mask, irc_raw=line,
-            irc_prefix=mask, irc_command=command, irc_paramlist=command_params, irc_ctcp_text=ctcp_text
+            bot=self.bot,
+            conn=self.conn,
+            event_type=event_type,
+            content_raw=content_raw,
+            content=content,
+            target=target,
+            channel=channel,
+            nick=nick,
+            user=user,
+            host=host,
+            mask=mask,
+            irc_raw=line,
+            irc_prefix=mask,
+            irc_command=command,
+            irc_paramlist=command_params,
+            irc_ctcp_text=ctcp_text,
         )
         return event