+++ tests/core_tests/irc_client_test.py
@@ -1,9 +1,15 @@
 import asyncio
+import logging
+import socket
 from asyncio import Task
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, call, patch
 
-from cloudbot.clients.irc import _IrcProtocol
+import pytest
+
+import cloudbot.clients.irc as irc
+from cloudbot.client import ClientConnectError
 from cloudbot.event import EventType
+from cloudbot.util import async_util
 
 
 class TestLineParsing:
@@ -20,7 +26,7 @@ class TestLineParsing:
 
         assert out == [
             {
-                "chan": None,
+                "chan": "server.host",
                 "content": None,
                 "content_raw": None,
                 "db": None,
@@ -69,7 +75,7 @@ class TestLineParsing:
             out.append(self._filter_event(e))
 
         conn.bot.process = func
-        proto = _IrcProtocol(conn)
+        proto = irc._IrcProtocol(conn)
         return conn, out, proto
 
     def test_broken_line_doesnt_interrupt(self):
@@ -82,7 +88,7 @@ class TestLineParsing:
 
         assert out == [
             {
-                "chan": None,
+                "chan": "server.host",
                 "content": None,
                 "content_raw": None,
                 "db": None,
@@ -132,7 +138,7 @@ class TestLineParsing:
         event = proto.parse_line(":server.host COMMAND this is :a command")
 
         assert self._filter_event(event) == {
-            "chan": None,
+            "chan": "server.host",
             "content": None,
             "content_raw": None,
             "db": None,
@@ -299,3 +305,217 @@ class TestLineParsing:
             "type": EventType.message,
             "user": "user",
         }
+
+
+@pytest.fixture()
+def patch_logger():
+    with patch.object(irc.logger, "isEnabledFor") as enabled:
+        enabled.return_value = True
+        with patch.object(irc.logger, "_log") as mock_log:
+            yield mock_log
+
+
+class TestConnect:
+    async def make_client(self) -> irc.IrcClient:
+        bot = MagicMock(loop=asyncio.get_event_loop(), config={})
+        conn_config = {
+            "connection": {
+                "server": "host.invalid",
+                "timeout": 1,
+                "bind_addr": "127.0.0.1",
+                "bind_port": 0,
+            }
+        }
+        client = irc.IrcClient(bot, "irc", "testconn", "foo", config=conn_config)
+        client.active = True
+        return client
+
+    @pytest.mark.asyncio()
+    async def test_exc(self, patch_logger):
+        client = await self.make_client()
+        runs = 0
+
+        # noinspection PyUnusedLocal
+        async def connect(timeout):
+            nonlocal runs
+            if runs == 5:
+                return
+
+            runs += 1
+            raise socket.error("foo")
+
+        client.connect = connect
+        await client.try_connect()
+        assert patch_logger.mock_calls == [
+            call(
+                logging.INFO,
+                "[%s|permissions] Created permission manager for %s.",
+                ("testconn", "testconn"),
+            ),
+            call(
+                logging.INFO,
+                "[%s|permissions] Reloading permissions for %s.",
+                ("testconn", "testconn"),
+            ),
+            call(
+                logging.DEBUG,
+                "[%s|permissions] Group permissions: %s",
+                ("testconn", {}),
+            ),
+            call(logging.DEBUG, "[%s|permissions] Group users: %s", ("testconn", {})),
+            call(
+                logging.DEBUG, "[%s|permissions] Permission users: %s", ("testconn", {})
+            ),
+            call(
+                logging.ERROR,
+                "[%s] Error occurred while connecting to %s (%s)",
+                ("testconn", "host.invalid:6667", "OSError: foo"),
+            ),
+            call(
+                logging.ERROR,
+                "[%s] Error occurred while connecting to %s (%s)",
+                ("testconn", "host.invalid:6667", "OSError: foo"),
+            ),
+            call(
+                logging.ERROR,
+                "[%s] Error occurred while connecting to %s (%s)",
+                ("testconn", "host.invalid:6667", "OSError: foo"),
+            ),
+            call(
+                logging.ERROR,
+                "[%s] Error occurred while connecting to %s (%s)",
+                ("testconn", "host.invalid:6667", "OSError: foo"),
+            ),
+            call(
+                logging.ERROR,
+                "[%s] Error occurred while connecting to %s (%s)",
+                ("testconn", "host.invalid:6667", "OSError: foo"),
+            ),
+        ]
+
+    @pytest.mark.asyncio()
+    async def test_timeout_exc(self, patch_logger):
+        client = await self.make_client()
+        runs = 0
+
+        # noinspection PyUnusedLocal
+        async def connect(timeout):
+            nonlocal runs
+            if runs == 5:
+                return
+
+            runs += 1
+            raise TimeoutError("foo")
+
+        client.connect = connect
+        await client.try_connect()
+        assert patch_logger.mock_calls == [
+            call(
+                20,
+                "[%s|permissions] Created permission manager for %s.",
+                ("testconn", "testconn"),
+            ),
+            call(
+                20,
+                "[%s|permissions] Reloading permissions for %s.",
+                ("testconn", "testconn"),
+            ),
+            call(10, "[%s|permissions] Group permissions: %s", ("testconn", {})),
+            call(10, "[%s|permissions] Group users: %s", ("testconn", {})),
+            call(10, "[%s|permissions] Permission users: %s", ("testconn", {})),
+            call(
+                40,
+                "[%s] Timeout occurred while connecting to %s",
+                ("testconn", "host.invalid:6667"),
+            ),
+            call(
+                40,
+                "[%s] Timeout occurred while connecting to %s",
+                ("testconn", "host.invalid:6667"),
+            ),
+            call(
+                40,
+                "[%s] Timeout occurred while connecting to %s",
+                ("testconn", "host.invalid:6667"),
+            ),
+            call(
+                40,
+                "[%s] Timeout occurred while connecting to %s",
+                ("testconn", "host.invalid:6667"),
+            ),
+            call(
+                40,
+                "[%s] Timeout occurred while connecting to %s",
+                ("testconn", "host.invalid:6667"),
+            ),
+        ]
+
+    @pytest.mark.asyncio()
+    async def test_other_exc(self, patch_logger):
+        client = await self.make_client()
+
+        # noinspection PyUnusedLocal
+        async def connect(timeout):
+            raise Exception("foo")
+
+        client.connect = connect
+        with pytest.raises(ClientConnectError):
+            await client.try_connect()
+
+        assert patch_logger.mock_calls == [
+            call(
+                20,
+                "[%s|permissions] Created permission manager for %s.",
+                ("testconn", "testconn"),
+            ),
+            call(
+                20,
+                "[%s|permissions] Reloading permissions for %s.",
+                ("testconn", "testconn"),
+            ),
+            call(10, "[%s|permissions] Group permissions: %s", ("testconn", {})),
+            call(10, "[%s|permissions] Group users: %s", ("testconn", {})),
+            call(10, "[%s|permissions] Permission users: %s", ("testconn", {})),
+        ]
+
+    @pytest.mark.asyncio()
+    async def test_one_connect(self):
+        client = await self.make_client()
+
+        async def _connect(timeout):
+            await asyncio.sleep(timeout)
+
+        client._connect = _connect
+        with pytest.raises(
+            ValueError,
+            match="Attempted to connect while another connect attempt is happening",
+        ):
+            await asyncio.gather(client.connect(2), client.connect(0))
+
+    @pytest.mark.asyncio()
+    async def test_create_socket(self):
+        client = await self.make_client()
+        client.loop.create_connection = mock = MagicMock()
+        fut = asyncio.Future(loop=client.loop)
+        fut.set_result((None, None))
+        mock.return_value = fut
+
+        await client.connect()
+
+
+class TestSend:
+    @pytest.mark.asyncio()
+    async def test_send_sieve_error(self):
+        proto = irc._IrcProtocol(MagicMock(loop=asyncio.get_event_loop()))
+        proto._connected = True
+        proto._transport = MagicMock()
+        sieve = object()
+        proto.bot.plugin_manager.out_sieves = [sieve]
+        proto.bot.plugin_manager.internal_launch = launch = MagicMock()
+        fut = async_util.create_future(proto.loop)
+        fut.set_result((False, None))
+        launch.return_value = fut
+
+        await proto.send("PRIVMSG #foo bar")
+        assert len(launch.mock_calls) == 1
+        assert launch.mock_calls[0][1][0] is sieve