+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed handling of colons in core IRC parser
 - Fix FML random URL
 - Update tvdb.py to v3 TVDB API
+- Fix channel parameter handling in IRC client
 ### Removed
 - twitch.py removed due to outdated API and lack of maintainer